Improved typings for Subscriptions context and authorization
