+++ packages/graphql/src/schema/resolvers/mutation/update.ts
@@ -23,7 +23,7 @@ import type { Node } from "../../../classes";
 import { translateUpdate } from "../../../translate";
 import { execute } from "../../../utils";
 import { publishEventsToSubscriptionMechanism } from "../../subscriptions/publish-events-to-subscription-mechanism";
-import type { Neo4jGraphQLComposedContext } from "../wrapper";
+import type { Neo4jGraphQLComposedContext } from "../composition/wrap-query-and-mutation";
 import getNeo4jResolveTree from "../../../utils/get-neo4j-resolve-tree";
 import type { Neo4jGraphQLTranslationContext } from "../../../types/neo4j-graphql-translation-context";
