+++ packages/graphql/src/schema/resolvers/query/aggregate.ts
@@ -21,7 +21,7 @@ import type { GraphQLResolveInfo } from "graphql";
 import { execute } from "../../../utils";
 import type { Node } from "../../../classes";
 import { translateAggregate } from "../../../translate";
-import type { Neo4jGraphQLComposedContext } from "../wrapper";
+import type { Neo4jGraphQLComposedContext } from "../composition/wrap-query-and-mutation";
 import getNeo4jResolveTree from "../../../utils/get-neo4j-resolve-tree";
 import type { Neo4jGraphQLTranslationContext } from "../../../types/neo4j-graphql-translation-context";
