+++ packages/graphql/src/schema/resolvers/query/fulltext.ts
@@ -24,7 +24,7 @@ import { translateRead } from "../../../translate";
 import type { Node } from "../../../classes";
 import type { FulltextContext } from "../../../types";
 import Cypher from "@neo4j/cypher-builder";
-import type { Neo4jGraphQLComposedContext } from "../wrapper";
+import type { Neo4jGraphQLComposedContext } from "../composition/wrap-query-and-mutation";
 import type { Neo4jGraphQLTranslationContext } from "../../../types/neo4j-graphql-translation-context";
 import getNeo4jResolveTree from "../../../utils/get-neo4j-resolve-tree";
