+++ packages/graphql/src/schema/resolvers/subscriptions/subscribe.ts
@@ -23,13 +23,14 @@ import { Neo4jGraphQLError } from "../../../classes";
 import type Node from "../../../classes/Node";
 import type { NodeSubscriptionsEvent, RelationshipSubscriptionsEvent, SubscriptionsEvent } from "../../../types";
 import { filterAsyncIterator } from "./filter-async-iterator";
-import type { SubscriptionEventType, SubscriptionContext } from "./types";
+import type { SubscriptionEventType } from "./types";
 import { updateDiffFilter } from "./update-diff-filter";
 import { subscriptionWhere } from "./where/where";
 import { subscriptionAuthorization } from "./where/authorization";
 import type { GraphQLResolveInfo } from "graphql";
 import { checkAuthentication } from "./authentication/check-authentication";
 import { checkAuthenticationOnSelectionSet } from "./authentication/check-authentication-selection-set";
+import type { Neo4jGraphQLComposedSubscriptionsContext } from "../composition/wrap-subscription";
 
 export function subscriptionResolve(payload: [SubscriptionsEvent]): SubscriptionsEvent {
     if (!payload) {
@@ -56,7 +57,7 @@ export function generateSubscribeMethod({
     return (
         _root: any,
         args: SubscriptionArgs,
-        context: SubscriptionContext,
+        context: Neo4jGraphQLComposedSubscriptionsContext,
         resolveInfo: GraphQLResolveInfo
     ): AsyncIterator<[SubscriptionsEvent]> => {
         checkAuthenticationOnSelectionSet(resolveInfo, node, type, context);
@@ -69,7 +70,7 @@ export function generateSubscribeMethod({
 
         checkAuthentication({ authenticated: concreteEntity, operation: "SUBSCRIBE", context });
 
-        const iterable: AsyncIterableIterator<[SubscriptionsEvent]> = on(context.plugin.events, type);
+        const iterable: AsyncIterableIterator<[SubscriptionsEvent]> = on(context.subscriptionsEngine.events, type);
         if (["create", "update", "delete"].includes(type)) {
             return filterAsyncIterator<[SubscriptionsEvent]>(iterable, (data) => {
                 return (
