+++ packages/graphql/src/schema/resolvers/subscriptions/types.ts
@@ -17,18 +17,9 @@
  * limitations under the License.
  */
 
-import type { Neo4jGraphQLSubscriptionsEngine } from "../../../types";
 import type { JWTPayload } from "jose";
-import type { Neo4jGraphQLSchemaModel } from "../../../schema-model/Neo4jGraphQLSchemaModel";
 
 export type SubscriptionEventType = "create" | "update" | "delete" | "create_relationship" | "delete_relationship";
-
-export type SubscriptionContext = {
-    plugin: Neo4jGraphQLSubscriptionsEngine;
-    jwt?: JWTPayload;
-    jwtPayloadFieldsMap?: Map<string, string>;
-    schemaModel: Neo4jGraphQLSchemaModel;
-};
 
 export type SubscriptionConnectionContext = {
     connectionParams?: {
