+++ packages/graphql/src/types/neo4j-graphql-translation-context.ts
@@ -18,7 +18,7 @@
  */
 
 import type { ResolveTree } from "graphql-parse-resolve-info";
-import type { Neo4jGraphQLComposedContext } from "../schema/resolvers/wrapper";
+import type { Neo4jGraphQLComposedContext } from "../schema/resolvers/composition/wrap-query-and-mutation";
 
 /**
  * A small extension to {@link Neo4jGraphQLComposedContext}, adding the {@link resolveTree} field.
